Add files via upload
After about an hour, I still have no idea how to animate this coin in a way that doesn't look weird/forced. There is a colored version and a gold version to try. maybe make it rotate like the asteroid as it falls or or partially rotate back & forth if the GUI allows that sort of motion.